refactor(loft-taxi): разбил по директориям и почистил неиспользуемое
