+++ src/components/SideLogo.css
@@ -9,7 +9,7 @@
 }
 
 .side-logo__link {
-    background-image: url('./lofttaxi-biglogo.png');
+    background-image: url('../resources/lofttaxi-biglogo.png');
     background-size: 192px 228px;
     width: 192px;
     height: 228px;
